started creation CRUD: 1- show all
