--- libs/mongoose.js
@@ -1,4 +1,0 @@
-var mongoose = require('mongoose');
-mongoose.connect('mongodb://localhost/tododb');
-
-module.exports= mongoose;
